sshconfig: allow to skip "Ensure config for each host exist"
By setting sshconfig_private_key_file = "" it's possible to skip
the "Ensure config for each host exist" task. This way it's possible
to use the role to only write extra ssh config.

This is necessary if you initially set up the manager from the seed
and have to configure certain SSH parameters to reach the Git server.

Signed-off-by: Christian Berendt <berendt@osism.tech>