Restructure vxlan definitions
Currently the definition for vxlans is an array of mappings where the
index and the value of the name key form the name of the networkd unit's
file. Any change to the order of the array will result in a leaked unit
with the old index, which is unexpected from a user's perspective and
needs to be cleaned up.
The name already needs to be unique, since it is used as the interface's
name. Therefore the structure could also be a mapping using the name as
keys and filename, e.g.:

```
network_vxlan_interfaces:
  vxlan0:
    vni: ..
    local_ip: ..
    ..
  vxlan1:
    ...
```

This ensures unique interface names, which are also usable as filenames.
Changing an interface name would still leak a file with the old name,
but that would be expected when deleting resources from ansible.

Signed-off-by: Jan Horstmann <horstmann@osism.tech>